+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+- Added support for pipelining with `Exists`, `Within`, `Match` and `TypeOf` keywords [#90](https://github.com/BernieWhite/PSRule/issues/90)
+
 ## v0.3.0-B190224 (pre-release)
 
 - Export variables to improve authoring experience [#83](https://github.com/BernieWhite/PSRule/issues/83)